rtsp-server: Limit number of connecting clients
Previously only the number of threads in the GstRTSPThreadPool could be
used to limit the amount of work that GstRTSPServer would do. While this
number limits the amount of work being done in parallel it does not
limit the number of connecting clients. This patch adds a new property
that limits the number of connecting clients. If this number is reached
a client is never allowed to send its request and GstRTSPServer will
simply disconnect the client without sending any response as this would
constitute a denial of service attack.

Fixes https://bugzilla.gnome.org/show_bug.cgi?id=759777

Signed-off-by: Marc Leeman <m.leeman@televic.com>