aggregator: Add signal pad-drained to notify actual drained status of pad
In some complex use cases (such as dynamic add/remove composing
images), application might want to know when a stream is completely
consumed.
The issue is that EOS event cannot respect the consumed
status of a stream, because a buffer followed by the EOS might have
long duration and therefore the buffer will be alive until running time
of the other stream is reached to the buffer's running time.

Closes https://gitlab.freedesktop.org/gstreamer/gst-plugins-bad/issues/786