timeline: Handle seeks ourself ensuring each track receives the event
We currently do not support seeking one single track at a time and
in case each track is not linked to a sink, we need to make sure
all the tracks receives the seek event.

To allow that we handle seek events ourself and seek each track
we own ourself.

To avoid a race where the pipeline seeks each sink while the first seek
is being dispatched to each track (leading to seek failing on flushing
pads), we make sure to send flushes downstream ourself and drop any
dataflow during the seek operation until we get the right
flush_stop/segment the track.