camerabin2: preview: let appsink go synchronously to GST_STATE_PLAYING
This commit set the property "async" on the appsink. This allow it to go
to GST_STATE_PAUSED (and _PLAYING) synchronously without waiting for the
first buffer (in this case, the posted preview) to arrive.

This make preview processing faster for the first image, as the pipeline
doesn't have to do the state change from _PAUSED to _PLAYING before
posting the preview. Without this change, the state changing can make
'preview-image' posting happen later than the actual image capture
itself (as the element doing capture is usually in the _PLAYING state
already).