baseparse: Don't ignore repeated incoming PTS
If the input stream has PTS set, then previously baseparse
would only take a new timestamp when it changes, but it's
not clear why it does that.

When parsing video streams this behaviour makes it lose
timestamps from incoming video frames if there are duplicates
due to upstream timestamping problems.