baseparse: avoid bus error when no output frames
There are media with containers exposing streams but without no
output frames. In this case media processing will fail, even if
there are other streams that can be processed, in case of
playbin2, and fire an error in bus at EOS of all streams in
playbin3.

This patch demotes the error to a warning.