element: Deactivate pads when removing them from elements
This is symmetrict with gst_element_add_pad(), which is nowadays also
activating pads when they're added to elements in the PAUSED/PLAYING
state.

Nonetheless it's safer and recommended to deactivate pads separately to
ensure that it happens at the right time in case the element needs the
streaming thread to be shut down already before the pad is being
removed.