pad: Don't mark sticky events as received if new sticky events are pending
The caller will send sticky events again from the beginning, and marking
the event as received might cause it to skip forwarding the current
event. This can then cause a new peer of our pad to not get notified
about this event, leading to for example wrong caps being configured on
it while receiving buffers for different caps.