pad: Force harder sending of EOS after other sticky event errors
Even if e.g. a CAPS event was rejected, we previously enforced sending
an EOS event downstream to ensure that the pipeline can EOS cleanly. Due
to changes elsewhere this has to be forced a bit harder now, especially
we must ignore errors from sending sticky events in more places.