Revert "h264parser: Expose framerate even if fixed_frame_rate flag isn't set"
This reverts commit 43006110827bdf0a049befb78bf1011676374b82.

The assumption was that:

There is nothing in the spec that state that framerate is not valid in
that case.

However, when looking at the speci (section E.2.1), it clearly states that:

fixed_frame_rate_flag equal to 1 indicates that the temporal distance
between the HRD output times of any two consecutive pictures in output
order is constrained as follows. fixed_frame_rate_flag equal to 0
indicates that no such constraints apply to the temporal distance
between the HRD output times of any two consecutive pictures in output order.
When fixed_frame_rate_flag is not present, it shall be inferred to be equal to 0.

Moreover, the process for calculating DeltaTfiDivisor is clearly
conditioned to this flag being set to 1:

When fixed_frame_rate_flag is equal to 1 for a coded video sequence
containing picture n [..]

There is no provision for the value of DeltaTfiDivisor when
fixed_frame_rate_flag is 0.