avviddec: use a more conservative default number of threads
When doing per-frame multithreading, FFmpeg introduces a
delay of one frame per thread. On a machine with a high
number of cores, the default maximum of 16 can easily
lead to deadlocks in multiqueue.

The ideal situation would have multiqueue query a latency
downstream to add to its maximum, but that's not easily
implemented.

For now simply use a lower default maximum of 4, this should
usually still be enough for real-time playback purposes, and
can of course be set by the user through the max-threads
property.