event: New option for live flushing
When dealing with flushes and live sources (i.e. that return
GST_STATE_CHANGE_NO_PREROLL), the states of sinks would be lost resulting in
clocks not advancing until those sinks were prerolled. Sadly gstbin (and
gstpipeline) no longer have the information that there were no-preroll sources
in the pipeline and that therefore prerolling should prevent staying in PLAYING.

As a workaround for this, add the possibility for live sources to be able to
indicate that the FLUSH_START should *not* cause a lost of state.