kmssink: support atomic api
The modesetting and page flip have differents work flow
in atomic API. And I think in atomic time, you can't scanout
a buffer in DRM CRTC, the only usage for modesetting is updating
the resolution for the monitor.

I have made the probing work of the legacy way to use the cached
data, which is necessary for the atomic way, but I don't think
the legacy driver would change its properties during this work.
Then it would be right to do that.