audio aggregator: Remove never triggered condition
At this point, there is a buffer, which means that there is
some data left in it. But we've read all that is in there. So
we're certain that the next byte written from that buffer will
be after the current output offset.

Add a g_assert() just to be safe