avtp: Properly handle pipeline latency
Currently the "processing-deadline" property is used to configure a max
delay for the AVTP avtpbasepayload elements, however this falls over in
many scenarios resulting in audio/video sync issues.

The main times this obviously occurs is when a video file is demuxed and
the video AND audio streams are both broadcast from the same talker. The
AVTP listener pipelines for video and audio rendering almost always have
different pipeline latencies, and unless the talker knows this in
advance and compensates for it by increasing the processing-deadline
high enough then there is always bad sync.

This is actually even worse than that because the current configuration
has the implied assumption that all the listeners have matching
pipelines, which is obviously not going to be true in networks with
mismatched hardware.

So in order to address this properly this patch adds support to the
avtpbasepayload and avtpbasedepayload elements to listen for
GST_EVENT_LATENCY events from the pipeline and track this latency to
offset the timestamp when converting the AVTP timestamp to AVTP
presentation time when payloading and depayloading.

This also makes the "processing-deadline" property redundant for the
avtpbasepayload elements so it has been removed.