h265parser: weaken the check of lower_bit_rate_constraint_flag_set.
The general_lower_bit_rate_constraint_flag flag only controls the HbrFactor
of the bitstream, and so it only changes the limitation of the level and
tier. This should not be a critical error for decoding the streams if this
flag is not set correctly. And so when we judge the profile, we should only
consider it as a mismatch bit, not a critical flag that excludes some profile
candidate.

Fix: #949