videodecoder: Limit QoS skip time
When video is running late, and qos handling is enabled,
the decoder would skip at least 2*diff frames, where
diff is how late frames are arriving downstream. If there's
a timestamping bug or a clock glitch, this can lead to some
very large amounts of video being skipped. Make the
maximum length of frames that are skipped configurable, with
a 0.5 second default - so we'll always output at least 1 frame
and see if the downstream QoS problem has resolved.