multiudpsink: add socket-txtime option
This option enables synchronization of packets on the kernel / device
via the SO_TXTIME socket option. This allows granular, sub-millisecond
synchronization of packets that can't be achieved reliably in the userland.

When this option is enabled, udpsink disables synchronization in GstBaseSink
and also provides a clock, a GstSystemClock that uses CLOCK_TAI. Currently,
synchronization will not work properly if another clock is selected.