rtsp-stream: Forward RFC7273 publish clock from media to stream
Previously the pipeline was fully pre-rolled before the SDP was
generated but since a while it is only blocked on pad probes. This means
that the clock is not forwarded to all the elements yet and can't be
used for generating the SDP.

To solve this forward the clock directly from the media to the stream
after construction.