meson: sort list of plugin names in summary output
Collect plugins into a dict, along with the plugin name
(which also means we no longer have to strip the gst prefix).

The plugin dict's keys() are returned in sorted order.

For now we still turn the dict into a list again for backwards
compat since the variable is used by the monorepo super project
and in the docs/ subdir, but we can change that later if we want to.