adaptivedemux2: No longer "Fixup" download duration
This was related to the way things worked before the refactoring of the code to
use the download helper. Also this issue is fixed with the previous commit which
now properly gets the *real* moment we get the first bytes back from the
server (i.e. when we get the headers).

This correction would also cause completely bogus bitrate estimates for
high-latency servers where getting the first byte would take some time but the
actual bandwith was sufficient.