qtmoovrecover: recover mp4 files larger than 4GB
- the remaining data size of mdat is declared as guint but the mdat_size
field is a guint64 so if the mdat_size is greater than 4GB the computed
remaining size is wrong because it is a 32 bit unsigned int and it overflows.
this patch declares the remaining data size as guint64 and, in case the
mp4 is larger than 4GB, it rewrites the mdat header using the extended
version.
- the error messages in case there are remaining bytes in the headers
or in the original mdat are inverted.
this patch fix the error messages.
fixes issue #976