rtpmanagerbad: use RTCP sink socket iso RTCP src
rtpsink was re-using the udpsrc socket to send and receive RTCP packets.
When an application is changing the pipeline state from NULL -> PLAYING
-> NULL -> PLAYING... udpsrc will close the socket that is shared with
udpsink. As a result, only the first time the pipeline goes to playing;
will it send RTCP packets, all other times, it will complain about the
closed socket (but send RTP).

By re-using the udpsink socket instead of the udpsrc socket and setting
the the shared socket to NULL before the state transition is applied on
udpsrc, the socket can be re-used in future transitions.