mpegtspacketizer: fix packet offset calculation after seek
After a seek, the packetizer has to sync itself. It does so internally
by updating its map_offset field (used when reading from its internal
buffer), but it must also update its offset field (used when setting
an offset on the packets it outputs).

With this change, the offset field on any given packet is identical
when reaching it sequentially (reading from the start), or when reaching
it after a seek.