mpegtspacketizer: make PTS calculation from offset more consistent
Converting from PCR time to GST time incurs a rounding error.

When calculating the PTS for a given packet, its raw PTS won't change,
but the PCR and PCR offset of the current group might, depending
for instance on whether the PTS is being calculated after a seek.

Converting those from PCR time to GST time before performing the
PTS calculation can therefore result in small (nanosecond magnitude)
differences for the same packet depending on the context.

Address this by performing the calculation in the PCR domain first, and
then only converting to the GST domain.