tsdemux: fix consistency of timestamps after seeks
By design, tsdemux will set `start` on its initial segment to the
earliest PTS of the streams in its selected program.

In order to get fully accurate seeks, that means we need to offset
the requested position by the same amount, for instance:

- `tsdemux` handles a set of streams with earliest PTS 1 second

- It pushes a segment with start = 1 second

- It then gets a request to seek at 1 second: before this patch,
  the target of the seek was the earliest PTS of the stream, thus
  resulting in a one-second inaccuracy on seek

This fixes the issue by offseting every seek request by the initial
offset, we already discard seeks before the first new segment is
calculated and sent out, and as this patch stores the initial offset
at that point no extra check is needed.