videoconvert: fix default gamma-mode value
While conversion from one gamma transfer function to another
is an expensive operation, the resulting image can be substantially
different, and we shouldn't pretend that the operation has taken
place by default.

We do have methods to avoid the conversion when the transfer
functions are "equivalent", and users can always set gamma-mode
to none if they know what they're doing or like being lied to.