baseparse: fix calculating the offset based on seek pos
If the accurate flag is enabled,Offset for the seek is calculated based on the last entry in gstIndex.
The seek position is modifed based on the offset.If there is a major difference between the seek position and ts of last GSTIndex entry leads to the delay
in resuming the playback after the seek.