d3d11videosink: Remove a feature for rendering on application's texture
The feature was added to support zero-copy rendering to
an application's texture when GstD3D11 was not a public library,
but that's a kind of hack.

Now GstD3D11 is a public library and user can convert Direct3D11
texture generated by GStreamer to application's own texture
using generic approach (appsink and GstD3D11Converter object).

Since the old feature is being more and more maintenance burden,
and there is generic/better approach which can be used by application now,
this commit removes the old hack.