mssdemux: Append live fragments ahead of manifest on manifest update
After manifest update, new list of fragments is built from scratch
and the old list is discarded. Losing any fragments that were ahead
of manifest, if they're still ahead of the new one.

If current position was over one of the lost fragments,
gst_mss_stream_seek won't be able to find the correct fragment,
falling back to fragment #0