webrtc: nice: Add support for unescaped turn server url
1. Escape userinfo of turn server url string from user before calling
gst_uri_from_string_escaped().

2. To use 'time-limited TURN credentials', a form of 'timestamp:username'
must be used as the username part of the TURN credentials.
Previously, to meet the turn server uri form of webrtcbin, the first ':'
of 'timestamp:username' should be escaped by user to avoid invalid parsing
that treats 'timestamp' as a username. This patch also checks the timestamp
existence and set the 'timestamp:username' as a username properly.