matroskademux: don't consider subrip to be pango-markup
matroskademux element was considering SubRip subtitles format
as raw pango-markup, which is not true: SubRip may have some tags
that are not supported by pango.
That is why to output the subtitles correctly we have to connect
subparse element downstream that would convert the buffers to pango.
Currently subparse simply erases the tags it doesn't know, but
still it seems to be the only correct way for the future: tags that
aren't supported yet can be added later