Fix deprecation warning in hotdoc.generate_doc() with Meson >= 0.64.1
Meson >= 0.64.0 started to validate type of objects passed to the
`dependencies:` kwarg and GStreamer was passing a few invalid args that
should have been in `depends:` instead.

`dependencies:` is for libraries that the doc generator needs to link
against. It doesn't make much sense to pass things like a gir target or
an executable to it.

`depends:` is for any type of target that needs to be built before we
generate the doc. That should be used for gir, the doc cache, etc.

Unfortunately, the new `depends:` kwarg has only been added in Meson
0.64.1.