GL/SDL examples: Link to SDL2main.lib on Visual Studio-like
Sadly, it is not enough for us to just link to SDL2.lib on Visual
Studio-like builds, as this is what the SDL2 pkg-config file will give
us from its CMake builds.

Since Meson's support for SDL2 only covers pkg-config and sdl2-config
but not yet CMake, look for SDL2main.lib when SDL2 is found.  We may be
able to drop this when Meson's SDL2 support covers CMake files as well,
by using the 'components:' flag in the dependency... call.