GL/SDL Example (sdlshare.c): Work around wglShareList() limitation
This uses the code more or less like the QML GL plugin so that we won't
have a current WGL context in another thread so that we can avoid the
issue where wglShareLists() (or so) fails as the device or resource is
busy (error code 0xaa).

The shader here, though, doesn't compile on Windows for at least some
drivers, but that would be a separate issue or commit.