kmssink: Remove unneeded crop handling
This bit of code calculated exactly the same thing that gst_kms_sink_set_caps
has already calculated into the sink size (GST_VIDEO_SINK_WIDTH and
GST_VIDEO_SINK_HEIGHT), which is stored into `src` a bit further up, so no need
to repeat this calculation.

At first glance the calculation might seem different, since this code would take
cropping into account, but the video size passed to set_caps would also already
be cropped, resulting in exactly the same values. This can be illustrated by
trying different cropped inputs and seeing that the resulting resolution is the
same:

With the original code:

```
$ GST_DEBUG=2,kmssink:7 gst-launch-1.0 -e videotestsrc ! video/x-raw, width=1360, height=768, pixel-aspect-ratio=1/1 ! videocrop bottom=100 ! kmssink connector-id=308 |& grep drmModeSetPlane | head -n 1
0:00:00.037172215 204347 0x56155789f800 TRACE                kmssink gstkmssink.c:1980:gst_kms_sink_show_frame:<kmssink0> drmModeSetPlane at (0,198) 2256x1108 sourcing at (0,0) 1360x668
$ GST_DEBUG=2,kmssink:7 gst-launch-1.0 -e videotestsrc ! video/x-raw, width=1360, height=768, pixel-aspect-ratio=2/1 ! videocrop bottom=100 ! kmssink connector-id=308 |& grep drmModeSetPlane | head -n 1
0:00:00.019149132 204393 0x560d741a1400 TRACE                kmssink gstkmssink.c:1980:gst_kms_sink_show_frame:<kmssink0> drmModeSetPlane at (0,475) 2256x554 sourcing at (0,0) 1360x668
$ GST_DEBUG=2,kmssink:7 gst-launch-1.0 -e videotestsrc ! video/x-raw, width=1360, height=768, pixel-aspect-ratio=1/2 ! videocrop bottom=100 ! kmssink connector-id=308 |& grep drmModeSetPlane | head -n 1
0:00:00.017162549 204410 0x55c5f3844000 TRACE                kmssink gstkmssink.c:1980:gst_kms_sink_show_frame:<kmssink0> drmModeSetPlane at (362,0) 1531x1504 sourcing at (0,0) 1360x668
```

And with the modified code:
```
$ GST_DEBUG=2,kmssink:7 gst-launch-1.0 -e videotestsrc ! video/x-raw, width=1360, height=768, pixel-aspect-ratio=1/1 ! videocrop bottom=100 ! kmssink connector-id=308 |& grep drmModeSetPlane | head -n 1
0:00:00.017571846 204057 0x55e5ba9a9400 TRACE                kmssink gstkmssink.c:1968:gst_kms_sink_show_frame:<kmssink0> drmModeSetPlane at (0,198) 2256x1108 sourcing at (0,0) 1360x668
$ GST_DEBUG=2,kmssink:7 gst-launch-1.0 -e videotestsrc ! video/x-raw, width=1360, height=768, pixel-aspect-ratio=2/1 ! videocrop bottom=100 !  kmssink connector-id=308 |& grep drmModeSetPlane | head -n 1
0:00:00.017174118 204160 0x5583a2841400 TRACE                kmssink gstkmssink.c:1968:gst_kms_sink_show_frame:<kmssink0> drmModeSetPlane at (0,475) 2256x554 sourcing at (0,0) 1360x668
$ GST_DEBUG=2,kmssink:7 gst-launch-1.0 -e videotestsrc ! video/x-raw, width=1360, height=768, pixel-aspect-ratio=1/2 ! videocrop bottom=100 !  kmssink connector-id=308 |& grep drmModeSetPlane | head -n 1
0:00:00.018545265 204187 0x56290bd0c800 TRACE                kmssink gstkmssink.c:1968:gst_kms_sink_show_frame:<kmssink0> drmModeSetPlane at (362,0) 1531x1504 sourcing at (0,0) 1360x668
```

Note that the `src.w` and `src.h` variables modified by the removed code are
only used by `gst_video_sink_center_rect` a few lines down (since they are
overwritten a few lines further on), so any changes in behavior should be
visible in `result` (the output of of `gst_video_sink_center_rect`) and thus be
visible in above debug output.

Looking in the git history, this code was introduced in 2016 in 37c670e235
(kmssink: Fix selection of source region) as part of a fix for
https://bugzilla.gnome.org/show_bug.cgi?id=767422. It seems that this particular
bit of code might not have been necessary back then, though maybe cropping
worked differently back then maybe. The testcase from the bug report
(pixel-aspect-ratio mismatch) has been broken again since then (both before and
this commit), but this is the subject of
https://gitlab.freedesktop.org/gstreamer/gstreamer/-/issues/1555 and will be
fixed in a subsequent commit.