kmssink: Fix pixel-aspect-ratio scaling
When the pixel-aspect-ratio of the video and display mismatch (and
can_scale=true), the sink size is set to a (slightly) scaled version of the
video size that compensates for this pixel-aspect-ratio difference.

However, when rendering, this corrected sink size was used as the source
rectangle, which could cause unintended cropping (when the size was smaller) or
rendering failure (when the size was larger). In the latter case, kmssink would
fallback to unscaled pixel-for-pixel playback which could cause unintended
(lack of) scaling.

This behavior was broken in commit 671d9fed3a (kmssink: Do not source using
padded width/height), which switched from taking the source size from vinfo to
the sink size. This commit now essentially reverts the change from that commit,
so this needs further improvement.

This fixes #1555.