h264parse: Add a mode to drop frames of specified type
In some cases one might want to ensure that some frames are not decoded,
for example to avoid unnecessary processing.

A similar behavior can be achieved using seeks with tricks modes but
that doesn't play well with live streams so adding a mode in `h264parse`
is more appropriate for that use case.