videoaggregator: Prefer video info from the largest input
Previously, we picked the video info of the last input the zorder-sorted sinkpad
list as the output video info. This could lead to some awkward results, such as
picking colorimetry/chroma-site values of a SD input and using that by default
for an HD output. It makes more sense to use the video info of the largest
input, ideally leading to less work in conversion and a consistent output
format.