v4l2object: ignore colorimetry addition in v4l2 plugin sink caps
When run pipeline with v4l2videodecoder and v4l2convert, decoder
and parser/demux negotiate a specific colorimetry which is not included in
video-color colorimertry list. when specific colorimetry negotiate with v4l2convert
sink caps which not include the colorimetry, and it will casue the caps
negotiation to fail.