validate: fix imprecision in test.mse.segment_future_past_nomse
In gst_qtdemux_stream_segment_get_boundaries, multiplying with the
segment rate would cause the result to be off by almost 100 ns, since
the value is too big to be represented precisely by a double. With
rational time, the numerator is much smaller, so the result will be more
precise and the test can be updated to reflect that.