h264parse: improve PPS parse error handling
Don't use the PPS if the referenced SPS was not found
(GST_H264_PARSER_BROKEN_LINK). It's not usable and the didn't finish
parsing it anyways, so it does not contain any useful data.

Instead fallback to the previous PPS and update parser state and header
flag accordingly.

This is similar gst-plugins-bad!59 but for PPS instead of SPS.