ci: Check the plugins of the builds
There have been times where a plugin stopped or started building
after a change without noticing, since they are all autofeatures.
This also has happend after updating the CI image.

This is especially problematic when they go missing, as the CI
is no longer testing them.

This commit adds the output of gst-inspect on fedora and windows
jobs under ci/data/{fedora, windows} and checks against them
for each build. If there's a diff, it creates a .diff as job
artifact and outputs a curl command to apply it (much like the
documentation job does)

This is very similar to the cerbero MR:

https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/813